Bug Fixes and minor refactors to Windows bootstrapping
Add chef_version to the node_resource. Add tests for chef_version, ssh, and winrm.
Move logic for chef_version into windows_template_binding and out of the actual template file.
Add a new method to the winrm/worker class. _upload_command will upload the command to the host if the command is too long to execute.