Turn off real key generation in ChefZero
When this flag is set to false, when you create a client or user, or regenerate their private key, chef-zero will always return you the exact same private key. The public key that goes with it sits in chef_zero.rb alongside the private key (they are consts).

When generate_real_keys is true (presently the default), chef-zero will generate a real private/public key, which is a very slow process (and rather useless for chef-zero).